requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
cryptography>=42.0.0
aiofiles>=23.2.1
pymongo>=3.12

openai>=1.12.0
google-generativeai>=0.4.1

google-api-python-client
google-auth-httplib2
google-auth-oauthlib
